Create directory structure for setuptools, split CFFI stuff out
We're going to need to be able to build this thing properly, and
setuptools can sorta do it. These notes[1] suggest that there's not yet
any "correct" way to build and package CFFI interfaces. It should be
something like this:

User -> import marquise (.py) -> CFFI shim (.so) -> libmarquise

CFFI compiles on import, which is no good for distribution. For now I've
pushed the CFFI stuff into a separate module (marquise_cffi), which can
be built into a similarly named .so module. The python module
(`marquise`) can then import that and it looks like things will work as
expected. There'll be a little more work to do to ensure that the
packaged python stuff only includes marquise.py, as marquise_cffi.py
shouldn't be needed.

[1]: https://bitbucket.org/cffi/cffi/issue/109/enable-sane-packaging-for-cffi