send_{simple,extended} should raise an exception if the write fails
This came to light while writing the testsuite. update_source throws an
exception if the write fails (eg. due to the spool file being
read-only), why shouldn't send_simple and send_extended?

I originally had them return True/False for success, but checking
return codes is kinda lame in my opinion (this may or may not be a
religious question, meh). If we want to stick with returning bools then
the test will need to be updated from `raises` to `assert`.