Fix string formatting of `address` in debug output
There's no guarantees that address is a number, the test coverage suite
showed this manifesting when certain expected breakage wasn't occurring.

Explicitly expecting a number (%d) was causing an exception to be thrown
early, meaning we don't hit the case where bad input is fed to
marquise_update_source() in the C code. Print as a string (%s) instead.

I've left other instances of %d intact, where the value is guaranteed to
only ever be an int (FFI.errno and the C return values).