Add GLib include_dirs to FFI.verify() call (brittle)
This is not ideal because it depends on where the GLib headers live on
the build host. I'm not 100% sure that I'm doing all this correctly, but
it seems to be right. Of course I bet I'll get an annoying surprise the
moment I try to build this anywhere other than my workstation.