Derive correct include_dirs to pass to FFI.verify()
The GLib header file glibconfig.h isn't in the same location on Debian
and Fedora, and God knows where it might be on other distros. We'll do
our best to find it and add the directory to the include path
accordingly. Doing it in this way should also be easily extensible if
the need ever arises.